Correccion calculo de nota.

+++ src/main/java/ar/edu/unq/virtuaula/model/Lesson.java
@@ -50,15 +50,15 @@ public class Lesson implements Serializable {
         return this.tasks.isEmpty() ? 0 : completed * 100 / this.tasks.size();
     }
 
-    public int qualification() {
-        int completed = (int) this.tasks.stream()
+    public double qualification() {
+        double completed = (double) this.tasks.stream()
                 .filter(task -> State.COMPLETED.equals(task.getState()) && task.getCorrectAnswer().equals(task.getAnswer()))
                 .mapToDouble(taskAnser -> taskAnser.getScore())
                 .sum();
         return this.tasks.isEmpty() ? 0 : validate(completed);
     }
 
-    private int validate(int completed) {
+    private double validate(double completed) {
         Assert.isTrue(this.maxNote >= completed, "The qualification exceed the max note.");
         return completed;
     }
